requirements.txt
streamlit
pandas
aiohttp
akamai-edgeauth
